+++ domain/src/main/kotlin/team/dahaeng/android/domain/community/model/Post.kt
@@ -1,4 +1,4 @@
-package team.dahaeng.android.domain.model
+package team.dahaeng.android.domain.community.model
 
 data class Post(
     val title: String,
